+++ db/migrate/20150414054139_add_name_to_mokio_users.rb
@@ -1,0 +1,6 @@
+class AddNameToMokioUsers < ActiveRecord::Migration
+  def change
+    add_column :mokio_users, :first_name, :string, after: :email
+    add_column :mokio_users, :last_name, :string, after: :first_name
+  end
+end